add custom check permissions and authorization

+++ src/main/java/com/example/retmix/controller/ProductController.java
@@ -1,7 +1,9 @@
 package com.example.retmix.controller;
 
 import com.example.retmix.dto.products.CreateOrUpdateProductDTO;
+import com.example.retmix.models.enums.AvailablePermission;
 import com.example.retmix.services.ProductService;
+import com.example.retmix.services.UserService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,9 +15,11 @@ import java.util.Map;
 public class ProductController {
 
     private final ProductService productService;
+    private final UserService userService;
 
-    public ProductController(ProductService productService) {
+    public ProductController(ProductService productService, UserService userService) {
         this.productService = productService;
+        this.userService = userService;
     }
 
     @GetMapping
@@ -24,18 +28,21 @@ public class ProductController {
     }
 
     @PostMapping
-    public ResponseEntity<?> store(@Valid @RequestBody CreateOrUpdateProductDTO createOrUpdateProductDTO) {
+    public ResponseEntity<?> store(@RequestHeader("authorization") String token, @Valid @RequestBody CreateOrUpdateProductDTO createOrUpdateProductDTO) {
+        userService.checkPermission(userService.getUserByToken(token), AvailablePermission.ADD_PRODUCT);
         return ResponseEntity.status(201).body(Map.of("data", productService.createProduct(createOrUpdateProductDTO)));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable int id) {
+    public ResponseEntity<?> delete(@RequestHeader("authorization") String token, @PathVariable int id) {
+        userService.checkPermission(userService.getUserByToken(token), AvailablePermission.REMOVE_PRODUCT);
         productService.removeProduct(id);
         return ResponseEntity.ok(Map.of("data", Map.of("message", "Продукт удален")));
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<?> update(@PathVariable int id, @Valid @RequestBody CreateOrUpdateProductDTO updateProductDTO) {
+    public ResponseEntity<?> update(@RequestHeader("authorization") String token, @PathVariable int id, @Valid @RequestBody CreateOrUpdateProductDTO updateProductDTO) {
+        userService.checkPermission(userService.getUserByToken(token), AvailablePermission.EDIT_PRODUCT);
         return ResponseEntity.ok(Map.of("data", productService.updateProduct(id, updateProductDTO)));
     }
 }
